requirements.txt
streamlit
fastapi
uvicorn
transformers
neo4j
python-dotenv
requests